Run alarmPanel.py each __init__                     Ensured that alarmPanel.py is ran each init in order to generate fresh data.
