+++ Mikael_FLASK_Test/tutorial/flaskr/__init__.py
@@ -14,6 +14,16 @@ def run_fake_data_script():
     subprocess.run(["python", script_path], check=True)
     print("Generated Fake Data")
 
+def generate_alarm_data():
+    # Get the directory of the current file (__init__.py)
+    dir_path = os.path.dirname(os.path.realpath(__file__))
+    
+    # Construct the full path to the alarm panel script
+    script_path = os.path.join(dir_path, "alarmPanel.py")
+    
+    subprocess.run(["python", script_path], check=True)
+    print("Generated Alarm Data")
+
 def create_app(test_config=None):
     """Create and configure an instance of the Flask application."""
     app = Flask(__name__, instance_relative_config=True)
@@ -49,13 +59,13 @@ def create_app(test_config=None):
     from . import blog
     from . import maintenance
     from . import engineer
-    from . import operator
+    from . import operatorInit
 
     app.register_blueprint(auth.bp)
     app.register_blueprint(blog.bp)
     app.register_blueprint(maintenance.bp, url_prefix='/maintenance')
     app.register_blueprint(engineer.bp, url_prefix='/engineer')
-    app.register_blueprint(operator.bp, url_prefix='/operator')
+    app.register_blueprint(operatorInit.bp, url_prefix='/operatorInit')
 
     # make url_for('index') == url_for('blog.index')
     # in another app, you might define a separate main index here with
@@ -66,5 +76,6 @@ def create_app(test_config=None):
     # Call init_db to initialize the database
     with app.app_context():
         db.init_db()
+        generate_alarm_data()  # Add this line
 
     return app
